The user should be able to have the headers Inferred when reading a Delimited file (#3472)
Implements https://www.pivotaltracker.com/story/show/181986831

+++ std-bits/table/src/main/java/org/enso/table/parsing/BooleanParser.java
@@ -33,7 +33,7 @@ public class BooleanParser extends IncrementalDatatypeParser {
   }
 
   @Override
-  protected Builder makeBuilderWithCapacity(long capacity) {
-    return new BoolBuilder((int) capacity);
+  protected Builder makeBuilderWithCapacity(int capacity) {
+    return new BoolBuilder(capacity);
   }
 }
